requirements.txt
selenium>=4.21
python-dotenv>=1.0
schedule>=1.2
rich>=13.7
psycopg2-binary>=2.9
